Add functionality of the game
Created references to all the html elements inside
the javascript (container, input, resetBtn, span,
createBtn)
Added event listener to the createBtn button that
calls the createDivs function on button click.
The createDivs function first fetches the value from
input box. The container div is then cleared so that
the rows and columns are created on a clean container
to avoid overlap.
Conditional statement checks if the input.value is
a number and if it is between 0-100, if not the user
is alerted using the '.alert' class and by the removing
the '.hide' class from the alert icon.
If the input satisfies all the conditions then the
'.alert' class is removed and the '.hide' class is
added back to the alert icon in order to hide it.
A nested loop is then used to create the grid by
adding row divs and then adding column divs inside
those row divs.
An event listener is also added to the column divs that
listen for the mouseover event and when it occurs
the updateColor function is called which changes
the background-color of the column divs randomly.
An event listener was added to the resetBtn which
resets the background-color of all the column divs
back to white.